+++ Applications/Hermes/hermesflaskexample_rx.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import requests
 import base64
 import json
