+++ Applications/Hermes/hermesflaskexample_tx.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import requests
 import ConfigParser
 import os
